requirements.txt
aiohttp
fpl
python-dotenv
schedule
smtplib
email
